T&A: explain inavailability of corrections
For the time being, tests with fixed question set support corrections.
Tests with one of the other question set types – random selection from
test question pools, dynamic question set (continuous testing mode) – do
not. This is also documented in the Corrections section
<https://iliasdocuworld.qualitus.de/goto.php?target=pg_2055_123&client_id=iliasdocuworld>
in the ILIAS Author’s Guide.

Or course, not all authors read that manual, and wonder why the table in
the corrections view is empty for tests with e.g. random question sets.
Replace the table with an explanation if the test does not use a fixed
question set. This should clarify the situation for users.

See <https://mantis.ilias.de/view.php?id=17901> for the corresponding
issue report. When corrections for tests with other question set types
are added (as per <https://mantis.ilias.de/view.php?id=17901#c41222>),
this commit should be reverted.